refactor(eval): extract scenario/trial-env/transcript leaves; eval.js under 700
ARCH-3 (follow-through): extracted three cohesive leaves from eval.js — eval-scenario.js
(parse/list/find scenario + claim-type), eval-trial-env.js (trial dir + isolation setup),
eval-transcript.js (stream-json + action parsing). All 17 moved functions byte-identical;
acyclic (leaves import only utils; eval.js re-exports the moved surface so every importer —
eval-benchmark, eval-preflight, eval-command, tests — keeps resolving unchanged).

eval.js 1077 -> 635 (under the 700 hard limit); new modules scenario 244 / trial-env 148 /
transcript 134. Full npm test green; lint clean. Completes the engine-decomposition goal:
cli.js 657, learning.js 744, eval.js 635, eval-graders.js 230 — all within/at the limit.
