+++ scripts/lib/eval.js
@@ -8,17 +8,29 @@
  */
 
 const fs = require('node:fs');
-const os = require('node:os');
 const path = require('node:path');
-const {
-  execCommand,
-  ensureDir,
-  getTimestamp,
-  sanitizeFilename,
-  CLAUDE_MAX_BUFFER,
-} = require('./utils');
+const { execCommand, ensureDir, getTimestamp, CLAUDE_MAX_BUFFER } = require('./utils');
 const stats = require('./eval-stats');
 const graders = require('./eval-graders');
+const {
+  SCENARIOS_DIR,
+  normalizeClaimType,
+  inferClaimType,
+  compactDate,
+  parseEvalName,
+  resolveMaxTurns,
+  parseScenario,
+  listScenarios,
+  findScenario,
+} = require('./eval-scenario');
+const {
+  createTrialDir,
+  cleanupTrialDir,
+  runSetup,
+  writeIsolationSettings,
+  buildIsolationSettings,
+} = require('./eval-trial-env');
+const { parseStreamJsonOutput, parseActionsFromTranscript } = require('./eval-transcript');
 
 /**
  * Eval scenario parsed from a markdown file
@@ -62,462 +74,8 @@ const graders = require('./eval-graders');
  */
 
 const EVALS_DIR = 'evals';
-const SCENARIOS_DIR = path.join(EVALS_DIR, 'scenarios');
 const RESULTS_DIR = path.join(EVALS_DIR, 'results');
 const BENCHMARKS_DIR = path.join(EVALS_DIR, 'benchmarks');
-
-const CLAIM_TYPES = new Set([
-  'non-regression',
-  'discriminative-lift',
-  'self-improvement-smoke',
-  'infra',
-]);
-
-function normalizeClaimType(raw) {
-  if (!raw) return undefined;
-  const value = String(raw).trim().toLowerCase().replace(/_/g, '-');
-  const aliases = {
-    lift: 'discriminative-lift',
-    discriminative: 'discriminative-lift',
-    'self-improvement': 'self-improvement-smoke',
-    smoke: 'self-improvement-smoke',
-    'self-improvement/smoke': 'self-improvement-smoke',
-    harness: 'infra',
-    infrastructure: 'infra',
-    'infra/harness': 'infra',
-  };
-  return aliases[value] || (CLAIM_TYPES.has(value) ? value : undefined);
-}
-
-function inferClaimType(scenario = {}) {
-  const explicit = normalizeClaimType(scenario.claimType);
-  if (explicit) return explicit;
-
-  const identity = [scenario.name, scenario.target].filter(Boolean).join('\n').toLowerCase();
-
-  if (identity.includes('self-improvement') || identity.includes('optional-learning')) {
-    return 'self-improvement-smoke';
-  }
-  if (
-    identity.includes('harness') ||
-    identity.includes('plugin-dir') ||
-    identity.includes('sessionstart') ||
-    scenario.scope === 'agent'
-  ) {
-    return 'infra';
-  }
-  if (scenario.verdictPolicy === 'non-regression' || scenario.preflight === 'skip') {
-    return 'non-regression';
-  }
-  if (scenario.scope === 'skill' || scenario.scope === 'workflow') {
-    return 'discriminative-lift';
-  }
-  return 'infra';
-}
-
-/**
- * Extract compact YYYYMMDD date from an ISO timestamp.
- * Used as fallback directory name when runId is not provided.
- * @param {string} [isoTimestamp] - ISO timestamp (defaults to now)
- * @returns {string} Compact date (e.g., '20260320')
- */
-function compactDate(isoTimestamp) {
-  return (isoTimestamp || getTimestamp()).slice(0, 10).replace(/-/g, '');
-}
-
-/**
- * Parse an eval name into scenario name and condition.
- * Used by appendResult, saveTranscript, and loadResults to derive storage paths.
- * @param {string} evalName - Eval name (may include -baseline or -treatment suffix)
- * @returns {{ scenarioName: string, condition: string }}
- */
-function parseEvalName(evalName) {
-  const isBaseline = evalName.endsWith('-baseline');
-  const isTreatment = evalName.endsWith('-treatment');
-  const condition = isBaseline ? 'baseline' : isTreatment ? 'treatment' : 'results';
-  const scenarioName = sanitizeFilename(evalName.replace(/-(baseline|treatment)$/, ''));
-  return { scenarioName, condition };
-}
-
-/**
- * Resolve max-turns with priority: CLI > scenario > pluginDir default (10) > undefined.
- * @param {Object} opts
- * @param {number} [opts.maxTurns] - CLI-specified max turns
- * @param {number} [opts.scenarioMaxTurns] - Scenario ## Max Turns value
- * @param {string} [opts.pluginDir] - Plugin dir (triggers default of 10)
- * @returns {number|undefined} Resolved max turns, or undefined if none applies
- */
-function resolveMaxTurns(opts = {}) {
-  if (opts.maxTurns != null) return opts.maxTurns;
-  if (opts.scenarioMaxTurns != null) return opts.scenarioMaxTurns;
-  if (opts.pluginDir) return 10;
-  return undefined;
-}
-
-/**
- * Parse an eval scenario from a markdown file
- * @param {string} filePath - Path to scenario markdown file
- * @param {string} [projectRoot] - Project root for ${PROJECT_ROOT} expansion
- * @returns {EvalScenario} Parsed scenario
- */
-function parseScenario(filePath, projectRoot) {
-  const content = fs.readFileSync(filePath, 'utf8');
-  const sections = {};
-  let currentSection = null;
-  const lines = content.split('\n');
-
-  let insideFence = false;
-  let heredocMarker = null;
-  for (const line of lines) {
-    if (/^`{3,}/.test(line)) insideFence = !insideFence;
-
-    // Ignore ## headers inside heredocs (they contain arbitrary text)
-    if (!insideFence && !heredocMarker) {
-      const heredocMatch = line.match(/<<-?\s*['"]?(\w+)['"]?\s*$/);
-      if (heredocMatch) heredocMarker = heredocMatch[1];
-    } else if (heredocMarker && line.trim() === heredocMarker) {
-      heredocMarker = null;
-    }
-
-    const headerMatch = !insideFence && !heredocMarker && line.match(/^##\s+(.+)/);
-    if (headerMatch) {
-      currentSection = headerMatch[1].trim().toLowerCase();
-      sections[currentSection] = [];
-    } else if (currentSection) {
-      sections[currentSection].push(line);
-    }
-  }
-
-  const nameMatch = content.match(/^#\s+Eval:\s*(.+)/m);
-  const name = nameMatch ? nameMatch[1].trim() : path.basename(filePath, '.md');
-
-  // Assertion bullets can be markdown checkboxes (`- [ ]` / `- [x]`) for text
-  // assertions OR behavioral forms like `- [tool_called] Bash:cmd`. Strip ONLY
-  // the markdown checkbox shell; leave behavioral `[tool_*]` prefixes intact so
-  // downstream parseBehavioralAssertion can recognise them. An earlier strip
-  // pattern `[ x\w_]*` was greedy enough to eat `tool_called` too, which caused
-  // behavioral assertions to be silently reclassified as text and routed to the
-  // model grader.
-  const assertions = (sections.assertions || [])
-    .filter((line) => line.match(/^-\s*\[[^\]]*\]/))
-    .map((line) =>
-      line
-        .replace(/^-\s*\[[ xX]*\]\s*/, '')
-        .replace(/^-\s+/, '')
-        .trim(),
-    );
-
-  const section = (key) => (sections[key] || []).join('\n').trim();
-  const trialsRaw = section('trials');
-  const trials = trialsRaw ? parseInt(trialsRaw, 10) : undefined;
-  const version = section('version');
-  const target = section('target');
-  const preflight = section('preflight').toLowerCase();
-  const verdictPolicy = section('verdict policy').toLowerCase();
-  const claimType = normalizeClaimType(section('claim type'));
-
-  // Plugin Dir: resolve ${PROJECT_ROOT} or use absolute path
-  const pluginDirRaw = section('plugin dir');
-  let pluginDir;
-  if (pluginDirRaw) {
-    pluginDir = projectRoot
-      ? pluginDirRaw.replace(/\$\{PROJECT_ROOT\}/g, projectRoot)
-      : pluginDirRaw;
-  }
-
-  // Max Turns: parse as integer
-  const maxTurnsRaw = section('max turns');
-  const maxTurns = maxTurnsRaw ? parseInt(maxTurnsRaw, 10) : undefined;
-
-  return {
-    name,
-    scope: section('scope') || 'skill',
-    scenario: section('scenario'),
-    context: section('context'),
-    assertions,
-    grader: section('grader') || 'code',
-    graderConfig: section('grader config'),
-    setup: section('setup'),
-    ...(target ? { target } : {}),
-    ...(trials && !Number.isNaN(trials) ? { trials } : {}),
-    ...(version ? { version } : {}),
-    ...(preflight ? { preflight } : {}),
-    ...(verdictPolicy ? { verdictPolicy } : {}),
-    ...(claimType ? { claimType } : {}),
-    ...(pluginDir ? { pluginDir } : {}),
-    ...(maxTurns && !Number.isNaN(maxTurns) ? { maxTurns } : {}),
-  };
-}
-
-/**
- * List all eval scenarios
- * @param {string} projectRoot - Project root directory
- * @returns {string[]} List of scenario file paths
- */
-function listScenarios(projectRoot) {
-  const scenariosPath = path.join(projectRoot, SCENARIOS_DIR);
-  if (!fs.existsSync(scenariosPath)) {
-    return [];
-  }
-  return fs
-    .readdirSync(scenariosPath)
-    .filter((f) => f.endsWith('.md'))
-    .map((f) => path.join(scenariosPath, f));
-}
-
-/**
- * Find a scenario by name
- * @param {string} name - Scenario name to find
- * @param {string} projectRoot - Project root directory
- * @returns {EvalScenario|undefined} Parsed scenario, or undefined if not found
- */
-function findScenario(name, projectRoot) {
-  for (const f of listScenarios(projectRoot)) {
-    const s = parseScenario(f, projectRoot);
-    if (s.name === name) return s;
-  }
-  return undefined;
-}
-
-/**
- * Create an isolated trial directory under the project root.
- * Uses .eval-trials/ (gitignored) to avoid macOS permission dialogs
- * that occur when running from /var/folders/ temp directories.
- * @param {string} evalName - Eval name for prefix
- * @param {number} trialNumber - Trial number
- * @param {string} [projectRoot] - Project root directory
- * @returns {string} Absolute path to trial directory
- */
-function createTrialDir(evalName, trialNumber, projectRoot) {
-  const base = projectRoot
-    ? path.join(projectRoot, '.eval-trials')
-    : path.join(process.cwd(), '.eval-trials');
-  ensureDir(base);
-  const prefix = `${sanitizeFilename(evalName)}-t${trialNumber}-`;
-  const trialDir = fs.mkdtempSync(path.join(base, prefix));
-  initializeGitBoundary(trialDir);
-  return trialDir;
-}
-
-/**
- * Initialize a minimal git repository boundary in a trial directory.
- * This is best-effort and intentionally avoids spawning git so unit-test exec
- * mocks and eval process accounting are not affected.
- * @param {string} trialDir - Trial directory path
- */
-function initializeGitBoundary(trialDir) {
-  try {
-    const gitDir = path.join(trialDir, '.git');
-    fs.mkdirSync(path.join(gitDir, 'objects'), { recursive: true });
-    fs.mkdirSync(path.join(gitDir, 'refs', 'heads'), { recursive: true });
-    fs.mkdirSync(path.join(gitDir, 'refs', 'tags'), { recursive: true });
-    fs.writeFileSync(path.join(gitDir, 'HEAD'), 'ref: refs/heads/main\n');
-    fs.writeFileSync(
-      path.join(gitDir, 'config'),
-      '[core]\n\trepositoryformatversion = 0\n\tfilemode = true\n\tbare = false\n\tlogallrefupdates = true\n',
-    );
-  } catch {
-    /* silent — git boundary is best-effort */
-  }
-}
-
-/**
- * Clean up a trial directory. Only removes paths inside .eval-trials/ or os.tmpdir().
- * @param {string} [trialDir] - Path to trial directory
- */
-function cleanupTrialDir(trialDir) {
-  if (!trialDir) return;
-  const isEvalTrial = trialDir.includes('.eval-trials');
-  const isTmp = trialDir.startsWith(os.tmpdir());
-  if (!isEvalTrial && !isTmp) return;
-  try {
-    fs.rmSync(trialDir, { recursive: true, force: true });
-  } catch {
-    /* silent — trial dir cleanup is best-effort */
-  }
-}
-
-/**
- * Run a setup command in the trial directory.
- * Injects PROJECT_ROOT env var so setup can reference project files.
- * @param {string} setupCommand - Shell command to execute
- * @param {string} trialDir - Working directory for setup
- * @param {string} [projectRoot] - Project root for $PROJECT_ROOT env var
- */
-function runSetup(setupCommand, trialDir, projectRoot) {
-  const env = { ...process.env };
-  if (projectRoot) env.PROJECT_ROOT = projectRoot;
-  const { exitCode, stderr } = execCommand('sh', ['-c', setupCommand], {
-    cwd: trialDir,
-    timeout: 30000,
-    env,
-  });
-  if (exitCode !== 0) {
-    throw new Error(`Setup failed: ${stderr}`);
-  }
-}
-
-/**
- * Write isolation settings to a trial directory.
- * Disables all user plugins, excludes user/project CLAUDE.md files,
- * and initializes a git repo to create a project boundary (prevents
- * Claude Code from walking up to the parent project's CLAUDE.md and rules).
- * @param {string} trialDir - Path to trial temp directory
- * @param {string} [cachedSettings] - Pre-built settings JSON (avoids repeated plugin list calls)
- */
-function writeIsolationSettings(trialDir, cachedSettings) {
-  const claudeDir = path.join(trialDir, '.claude');
-  fs.mkdirSync(claudeDir, { recursive: true });
-  const settings = cachedSettings || buildIsolationSettings();
-  fs.writeFileSync(path.join(claudeDir, 'settings.json'), settings);
-}
-
-/**
- * Build isolation settings JSON string (cached for reuse across trials).
- * Queries installed plugins and generates settings to disable all of them.
- * @param {Object} [opts] - Options
- * @param {boolean} [opts.excludeClaudeMd=true] - Exclude CLAUDE.md/rules (full isolation).
- *   Set false for semi-isolation (plugin-dir mode) where the plugin needs project context.
- * @returns {string} JSON string for .claude/settings.json
- */
-function buildIsolationSettings({ excludeClaudeMd = true } = {}) {
-  const baseSettings = {
-    autoMemoryEnabled: false,
-    ...(excludeClaudeMd
-      ? { claudeMdExcludes: ['**/CLAUDE.md', '**/CLAUDE.local.md', '**/rules/**'] }
-      : {}),
-  };
-  try {
-    const { stdout, exitCode } = execCommand('claude', ['plugin', 'list', '--json'], {
-      timeout: 10000,
-    });
-    if (exitCode !== 0 || !stdout) return JSON.stringify(baseSettings);
-    const plugins = JSON.parse(stdout);
-    if (!Array.isArray(plugins)) return JSON.stringify(baseSettings);
-    const disabled = {};
-    for (const p of plugins) disabled[p.id] = false;
-    return JSON.stringify({ ...baseSettings, enabledPlugins: disabled });
-  } catch {
-    return JSON.stringify(baseSettings);
-  }
-}
-
-/**
- * Parse stream-json output from `claude -p --output-format stream-json --verbose`.
- * Extracts assistant messages with tool calls to build a rich transcript
- * showing the full conversation flow (text + tool use in order).
- * Also extracts usage metrics (tokens, duration) from the terminal result event.
- * @param {string} rawOutput - Raw JSONL output from stream-json
- * @returns {{ textResult: string, richTranscript: string, usage: { input_tokens: number|null, output_tokens: number|null, duration_ms: number|null } }}
- */
-function parseStreamJsonOutput(rawOutput) {
-  if (!rawOutput)
-    return {
-      textResult: '',
-      richTranscript: '',
-      usage: { input_tokens: null, output_tokens: null, duration_ms: null },
-    };
-
-  const lines = rawOutput.split('\n').filter((l) => l.trim());
-  const parts = [];
-  let textResult = '';
-  const usage = { input_tokens: null, output_tokens: null, duration_ms: null };
-
-  for (const line of lines) {
-    let event;
-    try {
-      event = JSON.parse(line);
-    } catch {
-      continue;
-    }
-
-    if (event.type === 'assistant' && event.message?.content) {
-      for (const block of event.message.content) {
-        if (block.type === 'text' && block.text) {
-          parts.push(`[Assistant] ${block.text}`);
-        } else if (block.type === 'tool_use') {
-          const inputSummary = summarizeToolInput(block.name, block.input);
-          parts.push(`[Tool: ${block.name}] ${inputSummary}`);
-        }
-      }
-    } else if (event.type === 'result') {
-      textResult = event.result || '';
-      if (event.usage && typeof event.usage === 'object') {
-        usage.input_tokens = event.usage.input_tokens ?? null;
-        usage.output_tokens = event.usage.output_tokens ?? null;
-      }
-      usage.duration_ms = event.duration_ms ?? null;
-    }
-  }
-
-  return { textResult, richTranscript: parts.join('\n\n'), usage };
-}
-
-/**
- * Parse a rich transcript into structured Action objects.
- * Rich transcripts are produced by parseStreamJsonOutput() and contain
- * blocks like "[Tool: Bash] $ ls" and "[Assistant] some text" separated
- * by double newlines.
- *
- * NOTE: This parses markers generated by parseStreamJsonOutput(), not raw
- * assistant text. The [Tool: ...] markers come from tool_use blocks in the
- * stream-json output, so false positives from assistant prose are unlikely.
- * A future improvement could extract actions directly from JSON events.
- *
- * @param {string} richTranscript - Rich transcript text
- * @returns {Array<{type: string, name?: string, args?: string, content?: string, index: number}>}
- */
-function parseActionsFromTranscript(richTranscript) {
-  if (!richTranscript) return [];
-  const blocks = richTranscript.split('\n\n').filter((b) => b.trim());
-  const actions = [];
-  for (const block of blocks) {
-    const toolMatch = block.match(/^\[Tool: ([^\]]+)\]\s*(.*)/);
-    if (toolMatch) {
-      const firstLine = toolMatch[2].split('\n')[0];
-      actions.push({ type: 'tool', name: toolMatch[1], args: firstLine, index: actions.length });
-      continue;
-    }
-    const textMatch = block.match(/^\[Assistant\]\s*([\s\S]*)/);
-    if (textMatch) {
-      actions.push({ type: 'text', content: textMatch[1].trim(), index: actions.length });
-    }
-  }
-  return actions;
-}
-
-/**
- * Summarize tool input for transcript readability.
- * Shows key details without overwhelming the transcript.
- * @param {string} toolName - Tool name (Write, Edit, Bash, etc.)
- * @param {Object} input - Tool input parameters
- * @returns {string} Human-readable summary
- */
-function summarizeToolInput(toolName, input) {
-  if (!input) return '';
-  switch (toolName) {
-    case 'Write': {
-      const content = input.content || '';
-      return `${input.file_path || ''}\n\`\`\`\n${content}\n\`\`\``;
-    }
-    case 'Edit': {
-      const old = input.old_string || '';
-      const rep = input.new_string || '';
-      const maxLen = 300;
-      return `${input.file_path || ''} (replace "${old.slice(0, maxLen)}${old.length > maxLen ? '...' : ''}" → "${rep.slice(0, maxLen)}${rep.length > maxLen ? '...' : ''}")`;
-    }
-    case 'Read':
-      return input.file_path || '';
-    case 'Bash':
-      return `$ ${input.command || ''}`;
-    case 'Glob':
-      return `pattern: ${input.pattern || ''}`;
-    case 'Grep':
-      return `pattern: ${input.pattern || ''} ${input.path ? `in ${input.path}` : ''}`;
-    default:
-      return JSON.stringify(input).slice(0, 200);
-  }
-}
 
 /**
  * Run a single eval trial by spawning a Claude session.
